Cumulative Algorithms

+++ src/P6_DoubleProblems.java
@@ -8,7 +8,7 @@ public class P6_DoubleProblems {
 
 	public static void main(String[] args) {
 		double gpa = 3.2;
-		if( gpa * 3 == 9.6 ) {
+		if(  (float)(gpa * 3) == 9.6f ) {
 			System.out.println( "You earned enough credts");
 		}
 	}
